Fix entry point issues and implement modular architecture
