+++ MODULAR_ARCHITECTURE.md
@@ -1,0 +1,100 @@
+# QCMD Modular Architecture
+
+## Overview
+
+QCMD has been refactored from a single monolithic file into a modular architecture. This document describes the new package structure, the rationale behind the modularization, and how to extend the system.
+
+## Package Structure
+
+The QCMD code is now organized into the following packages:
+
+```
+qcmd_cli/
+├── __init__.py               # Package version and metadata
+├── __main__.py               # Main entry point
+├── commands/                 # Command handling
+│   ├── __init__.py
+│   └── handler.py            # CLI argument parsing and execution
+├── config/                   # Configuration management
+│   ├── __init__.py
+│   └── settings.py           # Config loading, saving, and defaults
+├── core/                     # Core functionality
+│   ├── __init__.py
+│   ├── command_generator.py  # Shell command generation
+│   └── interactive_shell.py  # Interactive shell
+├── log_analysis/             # Log analysis features
+│   ├── __init__.py
+│   ├── analyzer.py           # Log content analysis
+│   ├── log_files.py          # Log file discovery and selection
+│   └── monitor.py            # Log file monitoring
+├── ui/                       # User interface
+│   ├── __init__.py
+│   └── display.py            # Colors, banners, progress bars
+└── utils/                    # Utility functions
+    ├── __init__.py
+    ├── history.py            # Command history management
+    ├── session.py            # Session management and tracking
+    └── system.py             # System status and monitoring
+```
+
+## Design Principles
+
+The modular architecture follows these principles:
+
+1. **Single Responsibility**: Each module has a clear, focused purpose
+2. **Encapsulation**: Implementation details are hidden within modules
+3. **Dependency Management**: Dependencies are clearly defined through imports
+4. **Extensibility**: New features can be added without modifying existing code
+5. **Testability**: Modules can be tested in isolation
+
+## Benefits of Modularization
+
+The modular architecture provides several advantages:
+
+- **Maintainability**: Smaller, focused files are easier to understand and maintain
+- **Collaboration**: Different developers can work on different modules
+- **Code Reuse**: Functionality can be reused across the application
+- **Testability**: Modules can be tested individually
+- **Documentation**: Clear structure makes the codebase self-documenting
+- **Extensibility**: New features can be added by creating new modules
+
+## Backward Compatibility
+
+For backward compatibility, we maintain `qcmd_cli/qcmd.py` which imports and re-exports all the functionality from the modular structure. This allows existing scripts and users to continue using the previous API without modifications.
+
+## Module Dependencies
+
+Here's a high-level overview of the dependencies between modules:
+
+- `commands/handler.py` depends on most other modules as it orchestrates the application
+- `core/` modules depend on `config/`, `ui/`, and `utils/`
+- `log_analysis/` depends on `config/`, `ui/`, and `core/`
+- `utils/` depends on `config/` and `ui/`
+- `ui/` and `config/` have minimal dependencies on other modules
+
+## Extending the System
+
+To add new functionality to QCMD:
+
+1. Identify the appropriate package for your feature
+2. Create a new module or extend an existing one
+3. Update the appropriate entry points in `commands/handler.py`
+4. Re-export the functionality in `qcmd.py` for backward compatibility
+
+## Testing
+
+The modular architecture makes testing easier:
+
+- Unit tests can focus on individual modules
+- Integration tests can verify interactions between modules
+- The test file `tests/test_modular_imports.py` verifies that all modules can be imported correctly
+
+## Future Improvements
+
+Potential areas for future development:
+
+1. Add more comprehensive test coverage for individual modules
+2. Introduce plugin architecture for custom commands and features
+3. Develop separate user interfaces (CLI, GUI, web) using the same core
+4. Enhance logging and debugging capabilities
+5. Add support for additional AI backends beyond Ollama
