+++ migrate_to_modular.py
@@ -1,0 +1,46 @@
+#!/usr/bin/env python3
+"""
+Migration helper for QCMD refactoring.
+
+This script outlines the steps needed to migrate from the single-file approach
+to the new modular package structure.
+"""
+
+def main():
+    """
+    Print migration steps to guide the refactoring process.
+    """
+    print("QCMD Modular Migration Guide")
+    print("============================\n")
+    
+    print("The modular structure is now set up with the following packages:")
+    print("1. qcmd_cli/core/ - Core command generation and execution")
+    print("2. qcmd_cli/ui/ - User interface and display functionality")
+    print("3. qcmd_cli/utils/ - Utility functions for history, sessions, etc.")
+    print("4. qcmd_cli/log_analysis/ - Log analysis functionality")
+    print("5. qcmd_cli/config/ - Configuration management")
+    print("6. qcmd_cli/commands/ - Command-line argument handling\n")
+    
+    print("Migration Steps:")
+    print("--------------")
+    print("1. Keep the original qcmd.py file as a reference until migration is complete")
+    print("2. For each module, copy the function implementations from qcmd.py to the appropriate file")
+    print("   - Update imports within each function to use the new module structure")
+    print("   - Ensure constants and class definitions are properly included")
+    print("3. Update pyproject.toml to ensure all packages are included")
+    print("4. Update entry points in pyproject.toml to use the modular structure")
+    print("5. Run tests to ensure the refactored code works correctly")
+    print("6. Once refactoring is complete, remove the original qcmd.py file\n")
+    
+    print("Example of how to migrate a function:")
+    print("-----------------------------------")
+    print("1. Find a function in the original qcmd.py file")
+    print("2. Identify which module it belongs to")
+    print("3. Copy the function implementation to the appropriate file")
+    print("4. Update any imports or dependencies to use the new module structure")
+    print("5. Repeat for all functions\n")
+    
+    print("Remember to keep the same function signatures to maintain compatibility.")
+
+if __name__ == "__main__":
+    main()
