+++ qcmd_cli/commands/__init__.py
@@ -1,0 +1,4 @@
+"""
+Command handling for QCMD CLI.
+This module contains command-line argument parsing and execution.
+"""
