+++ qcmd_cli/config/__init__.py
@@ -1,0 +1,4 @@
+"""
+Configuration management for QCMD CLI.
+This module contains configuration loading, saving, and default settings.
+"""
