+++ qcmd_cli/core/__init__.py
@@ -1,0 +1,4 @@
+"""
+Core functionality for QCMD CLI.
+This module contains the command generation and interactive shell functionality.
+"""
