+++ qcmd_cli/log_analysis/__init__.py
@@ -1,0 +1,4 @@
+"""
+Log analysis functionality for QCMD CLI.
+This module contains log file analysis, monitoring, and selection functions.
+"""
