+++ qcmd_cli/log_analysis/analyzer.py
@@ -1,0 +1,55 @@
+#!/usr/bin/env python3
+"""
+Log analysis functionality for QCMD.
+"""
+import os
+import re
+import json
+import time
+import signal
+from typing import List, Dict, Optional, Tuple, Any
+
+# Import from local modules once they are created
+# from ..config.settings import DEFAULT_MODEL
+
+# For now, define defaults here
+DEFAULT_MODEL = "llama3:latest"
+
+def analyze_log_file(log_file: str, model: str = DEFAULT_MODEL, background: bool = False, analyze: bool = True) -> None:
+    """
+    Analyze a log file using AI.
+    
+    Args:
+        log_file: Path to the log file
+        model: Model to use for analysis
+        background: Whether to run in background mode
+        analyze: Whether to perform analysis (vs just monitoring)
+    """
+    # Implementation will be moved from original qcmd.py
+    pass
+
+def analyze_log_content(log_content: str, log_file: str, model: str = DEFAULT_MODEL) -> None:
+    """
+    Analyze the content of a log file.
+    
+    Args:
+        log_content: Content of the log file
+        log_file: Path to the log file (for reference)
+        model: Model to use for analysis
+    """
+    # Implementation will be moved from original qcmd.py
+    pass
+
+def read_large_file(file_path: str, chunk_size: int = 1024 * 1024) -> str:
+    """
+    Read a large file efficiently.
+    
+    Args:
+        file_path: Path to the file
+        chunk_size: Size of each chunk to read
+        
+    Returns:
+        Content of the file as a string
+    """
+    # Implementation will be moved from original qcmd.py
+    pass
