+++ qcmd_cli/ui/__init__.py
@@ -1,0 +1,4 @@
+"""
+User interface components for QCMD CLI.
+This module contains display formatting, colors, and UI elements.
+"""
