+++ qcmd_cli/utils/__init__.py
@@ -1,0 +1,4 @@
+"""
+Utility functions for QCMD CLI.
+This module contains history, session, and system utility functions.
+"""
